requirements.txt
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
drf-spectacular==0.25.1
drf-yasg==1.21.4
email-validator==1.3.0
Jinja2==3.1.2
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
pymongo==3.12.3
python-decouple==3.6
python-dotenv==0.20.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
whitenoise==6.2.0
uwsgi>=2.0.20,<2.1
python-json-logger==2.0.4